+++ assets/block-editor/column/index.js
@@ -8,15 +8,13 @@
  * Import dependencies
  */
 import classnames from 'classnames';
-
 import extractClasses from '../../util/extractClasses';
-
-const { __ } = wp.i18n;
-const { addFilter } = wp.hooks;
-const { createHigherOrderComponent } = wp.compose;
-const { Fragment, useState } = wp.element;
-const { InspectorControls } = wp.blockEditor;
-const { FormTokenField, PanelBody } = wp.components;
+import { __ } from '@wordpress/i18n';
+import { addFilter } from '@wordpress/hooks';
+import { createHigherOrderComponent } from '@wordpress/compose';
+import { Fragment, useState } from '@wordpress/element';
+import { InspectorControls } from '@wordpress/block-editor';
+import { FormTokenField, PanelBody } from '@wordpress/components';
 
 /**
  * Identify the extension's name