+++ assets/block-library/grd/grd-item/index.js
@@ -17,9 +17,8 @@ import Icon from './icon.inline.svg';
 import edit from './grd-item-edit';
 import save from './grd-item-save';
 import block from './block.json';
-
-const { __ } = wp.i18n;
-const { registerBlockType } = wp.blocks;
+import { __ } from '@wordpress/i18n';
+import { registerBlockType } from '@wordpress/blocks';
 
 /**
  * Register block