+++ assets/util/block-ui/LinkSelector/index.js
@@ -11,12 +11,10 @@ const { link } = attributes; // { url: 'https://example.com', opensInNewTab: fal
 />
  */
 import { link as linkIcon } from '@wordpress/icons';
-
-const { __ } = wp.i18n;
-const { useState } = wp.element;
-const { BlockControls, __experimentalLinkControl: LinkControl } =
-	wp.blockEditor;
-const { ToolbarGroup, ToolbarButton, Popover, TextControl } = wp.components;
+import { __ } from '@wordpress/i18n';
+import { useState } from '@wordpress/element';
+import { BlockControls, __experimentalLinkControl as LinkControl } from '@wordpress/block-editor';
+import { ToolbarGroup, ToolbarButton, Popover, TextControl } from '@wordpress/components';
 
 /**
  * Link selector