+++ includes/cli/templates/block-inner/index.js
@@ -12,9 +12,8 @@ import Icon from './icon.inline.svg';
 import edit from './template-item-slug-edit';
 import save from './template-item-slug-save';
 import block from './block.json';
-
-const { __ } = wp.i18n;
-const { registerBlockType } = wp.blocks;
+import { __ } from '@wordpress/i18n';
+import { registerBlockType } from '@wordpress/blocks';
 
 /**
  * Register block