+++ includes/cli/templates/block/template-slug-save.js
@@ -11,8 +11,7 @@
  * Import save dependencies
  */
 import classnames from 'classnames';
-
-const { useBlockProps, useInnerBlocksProps, RichText } = wp.blockEditor;
+import { useBlockProps, useInnerBlocksProps, RichText } from '@wordpress/block-editor';
 
 /**
  * Generate block HTML to save to the database